+++ .gitignore
@@ -1,1 +1,3 @@
 *.pyc
+data.db
+node_modules/
