+++ ServerAndWebapp/package.json
@@ -6,17 +6,9 @@
     "start": "node ./bin/www"
   },
   "dependencies": {
-    "bluebird": "^3.4.6",
     "body-parser": "~1.15.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.2.0",
     "ejs": "^2.5.2",
     "express": "~4.13.4",
-    "mongoose": "^4.6.3",
-    "morgan": "~1.7.0",
-    "request": "^2.75.0",
-    "request-promise": "^4.1.1",
-    "serve-favicon": "~2.3.0",
-    "underscore": "^1.8.3"
+    "sqlite3": "^3.1.6"
   }
 }
