--- ServerAndWebapp/public/app/views/login.html
@@ -1,26 +1,0 @@
-<div class="login-page"></div>
-<div class="clearfix"></div>
-<div class="wrapper-page">
-  <div id="logo" class="text-center m-t-10">IOTGarden</div>
-  <div class="card-box m-t-20">
-    <div class="panel-body">
-      <form class="form-horizontal m-t-20" ng-submit="login()">
-        <div class="form-group ">
-          <div class="col-xs-12">
-            <input class="form-control" type="email" required="" placeholder="Email" ng-model="user.email" autofocus>
-          </div>
-        </div>
-        <div class="form-group">
-          <div class="col-xs-12">
-            <input class="form-control" type="password" required="" placeholder="Password" ng-model="user.password">
-          </div>
-        </div>
-        <div class="form-group text-center m-t-30">
-          <div class="col-xs-12">
-            <button class="btn btn-success btn-bordred btn-block" type="submit">Log In</button>
-          </div>
-        </div>
-      </form>
-    </div>
-  </div>
-</div>
