--- ServerAndWebapp/services/index.js
@@ -1,3 +1,0 @@
-exports.user = require('./user.js');
-exports.garden = require('./garden.js');
-exports.sensor = require('./sensor.js');
