--- ServerAndWebapp/utils/index.js
@@ -1,2 +1,0 @@
-exports.response = require('./response.js');
-exports.auxMethods = require('./auxMethods.js');
