+++ ServerAndWebapp/views/index.ejs
@@ -1,12 +1,12 @@
 <!DOCTYPE html>
-<html ng-app="iotGarden">
+<html>
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="IOT Garden webapp">
     <meta name="author" content="@shenrgc">
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css">
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css">
     <link rel="stylesheet" href="stylesheets/style.css" type="text/css" />
     <title>IOT Garden</title>
@@ -16,29 +16,116 @@
         Please enable Javascript to experience this site :)
     </noscript>
     <div ng-controller="generalController">
-      <div id="content-page" ui-view></div>
+      <div id="content-page">
+
+		  <div class="dashboard-page"></div>
+		  <div class="navbar navbar-default" role="navigation">
+		    <div class="container">
+		      <ul class="nav navbar-nav navbar-left" style="margin-top: 10px;">
+		        <li>
+		          <h4>IOTGarden</h4>
+		        </li>
+		      </ul>
+		      <ul class="nav navbar-nav navbar-right">
+		        <li>
+		        </li>
+		      </ul>
+		    </div>
+		  </div>
+		  <div class="content">
+		    <div class="container">
+		      <!-- Charts info -->
+		      <div class="row">
+		        <div class="col-md-12">
+		          <div class="card-box">
+		            <h4 class="header-title">Controllers</h4>
+		            <div class="row">
+		              <div class="col-md-3 m-t-10 text-center">
+		                <h5 style="color: #428bca;"><b>Wind</b></h5>
+		                <div class="radio">
+		                  <label><input type="radio" name="fanopt" value="OFF">Off</label>
+		                  <label><input type="radio" name="fanopt" value="ON">On</label>
+		                  <label><input type="radio" name="fanopt" value="AUTO">Auto</label>
+						  <label><input type="radio" name="fanopt" value="SCHEDULE">Schedule</label>
+		                </div>
+
+		              </div>
+		              <div class="col-md-3 m-t-10 text-center">
+		                <h5 style="color: #f0ad4e;"><b>Light</b></h5>
+		                <div class="radio">
+		                  <label><input type="radio" name="lightopt" value="OFF">Off</label>
+		                  <label><input type="radio" name="lightopt" value="ON">On</label>
+		                  <label><input type="radio" name="lightopt" value="AUTO">Auto</label>
+						  <label><input type="radio" name="lightopt" value="SCHEDULE">Schedule</label>
+		                </div>
+		              </div>
+		              <div class="col-md-3 m-t-10 text-center">
+		                <h5 style="color: #5bc0de;"><b>Water</b></h5>
+		                <div class="radio">
+		                  <label><input type="radio" name="irrigateopt" value="Off">Off</label>
+		                  <label><input type="radio" name="irrigateopt" value="On">On</label>
+		                  <label><input type="radio" name="irrigateopt" value="Auto">Auto</label>
+						  <label><input type="radio" name="irrigateopt" value="SCHEDULE">Schedule</label>
+		                </div>
+		              </div>
+		            </div>
+		          </div>
+		        </div>
+		        <div class="col-md-12">
+		          <div class="card-box">
+		            <h4 class="header-title">Exterior humidity</h4>
+		            <div class="row">
+		              <div class="col-md-6 m-t-20" id="ext-humidity-chart1" style="height: 150px;"></div>
+		              <div class="col-md-6 m-t-20" id="ext-humidity-chart2" style="height: 150px;"></div>
+		            </div>
+		            <h4 class="header-title">Interior humidity</h4>
+		            <div class="row">
+		              <div class="col-md-6 m-t-20" id="int-humidity-chart1" style="height: 150px;"></div>
+		              <div class="col-md-6 m-t-20" id="int-humidity-chart2" style="height: 150px;"></div>
+		            </div>
+		          </div>
+		        </div>
+		        <div class="col-md-12">
+		          <div class="card-box">
+		            <h4 class="header-title">Exterior temperature</h4>
+		            <div class="row">
+		              <div class="col-md-6 m-t-20" id="ext-temperature-chart1" style="height: 150px;"></div>
+		              <div class="col-md-6 m-t-20" id="ext-temperature-chart2" style="height: 150px;"></div>
+		            </div>
+		            <h4 class="header-title">Interior temperature</h4>
+		            <div class="row">
+		              <div class="col-md-6 m-t-20" id="int-temperature-chart1" style="height: 150px;"></div>
+		              <div class="col-md-6 m-t-20" id="int-temperature-chart2" style="height: 150px;"></div>
+		            </div>
+		          </div>
+		        </div>
+		        <div class="col-md-12">
+		          <div class="card-box">
+		            <h4 class="header-title">Light percentage</h4>
+		            <div class="m-t-20" id="light-chart" style="height: 150px;"></div>
+		          </div>
+		        </div>
+		      </div>
+		    </div>
+		  </div>
+
+
+	  </div>
     </div>
     <!-- Libraries  -->
     <script src="https://code.jquery.com/jquery-3.0.0.min.js"></script>
+
+
     <script src="https://code.jquery.com/ui/1.11.4/jquery-ui.min.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.5.7/angular.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.3.1/angular-ui-router.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.5.7/angular-cookies.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-bootstrap/1.3.2/ui-bootstrap-tpls.min.js"></script>
+
     <!-- Plot -->
     <script src="scripts/jquery.flot.js"></script>
     <script src="scripts/jquery.flot.time.js"></script>
     <script src="scripts/jquery.flot.resize.js"></script>
     <script src="scripts/jquery.flot.tooltip.min.js"></script>
-    <!-- Configuration -->
-    <script src="app/main.js"></script>
-    <!-- Controllers -->
-    <script src="app/controllers/generalCtrl.js"></script>
-    <script src="app/controllers/loginCtrl.js"></script>
-    <script src="app/controllers/dashboardCtrl.js"></script>
-    <!-- Services -->
-    <script src="app/services/mainService.js"></script>
-    <script src="app/services/gardenService.js"></script>
+
+    <script src="scripts/dashboardCtrl.js"></script>
+
     <!-- End build -->
   </body>
 </html>
